+++ README.md
@@ -5,4 +5,4 @@ Código utilizado en el workshop *Programación funcional con Scala y aplicacion
 
 ### Descripción de la charla
 
-El objetivo del workshop es introducir las abstracciones y principios básicos de la programación funcional a programadores noveles en este paradigma, utilizando Scala como lenguaje de referencia y el framework de desarrollo Web Play. La idea es mostrar cómo los problemas a los que se enfrenta un desarrollador Web pueden ser resueltos mediante funciones "puras", tipos algebraicos, pattern matching, functores, mónadas, etc. Se utilizarán para ello ejemplos sencillos de la programación con Play sobre BodyParsers, Iteratees, AsynchResults, etc. El workshop se desarrollará en base a una metodología de live-coding de tal manera que todo aquel que quiera pueda seguir con su portátil los ejemplos presentados. En resumen, el workshop tiene un doble propósito: enseñar qué significa programar funcionalmente y cuales son sus ventajas, y mostrar cómo se aplica esta filosofía a problemas familiares para los asistentes como es el desarrollo Web. 
+El objetivo del workshop es introducir las abstracciones y principios básicos de la programación funcional a programadores noveles en este paradigma, utilizando Scala como lenguaje de referencia y el framework de desarrollo Web Play. La idea es mostrar cómo los problemas a los que se enfrenta un desarrollador Web pueden ser resueltos mediante funciones "puras", tipos algebraicos, pattern matching, functores, mónadas, etc. Se utilizarán para ello ejemplos sencillos de la programación con Play sobre BodyParsers, Iteratees, AsynchResults, etc. El workshop se desarrollará en base a una metodología de live-coding de tal manera que todo aquel que quiera pueda seguir con su portátil los ejemplos presentados. En resumen, el workshop tiene un doble propósito: enseñar qué significa programar funcionalmente y cuáles son sus ventajas, y mostrar cómo se aplica esta filosofía a problemas familiares para los asistentes como es el desarrollo Web.
