+++ app/controllers/DictionaryApp.scala
@@ -1,9 +1,9 @@
 package controllers
 
-import models._
+import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.Implicits.global
 
 import play.api._
-import play.api.cache.Cache
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.api.libs.iteratee._
 import play.api.libs.json._
@@ -11,45 +11,9 @@ import play.api.libs.ws._
 import play.api.mvc._
 import play.api.Play.current
 
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.Implicits.global
+import models._
 
 trait DictionaryApp { this: Controller =>
-
-  object Dictionary {
-
-    type Dictionary = Map[String, String]
-
-    private def dictionary = Cache.getOrElse[Dictionary]("dictionary")(Map())
-
-    def contains(word: String) = dictionary isDefinedAt word 
-
-    def get(word: String) = dictionary get word
-
-    def set(entry: (String, String)) = {
-      Logger.info(s"Adding word '${entry._1}' to dictionary")
-      Cache.set("dictionary", dictionary + entry)
-    }
-  }
-
-  object Users {
-
-    type Users = Map[String, User]
-
-    private def users = Cache.getOrElse[Users]("users")(Map())
-
-    def get(nick: String) = users get nick
-
-    def add(user: User) = {
-      Logger.info(s"Adding '${user.nick}' to user list.")
-      Cache.set("users", users + (user.nick -> user))
-    }
-
-    // adds some initial users, this could be moved to an external service
-    add(User("Mr", "Proper", Option(READ_WRITE)))
-    add(User("Don", "Limpio", Option(READ)))
-    add(User("Wipp", "Express"))
-  }
 
   val USER_HEADER_NAME = "user"
