+++ app/models/Permission.scala
@@ -16,4 +16,25 @@ object Permission {
     case Some(READ_WRITE) => true
     case _           => false
   }
+
+  // buen ejemplo de fold, y como ejercicio
+
+  val canRead2: User => Boolean = 
+    _.permission.fold(false){ _ => true }
+
+  val canWrite2: User => Boolean = 
+    _.permission.fold(false){ _ == READ_WRITE }
+    
+}
+
+// también se pueden definir como extractores. y utilizarse: user match { case CanRead => ... }
+
+object CanRead{
+  def unapply(user: User): Boolean = 
+    user.permission.fold(false){ _ => true }    
+}
+
+object CanWrite{
+  def unapply(user: User): Boolean = 
+    user.permission.fold(false){ _ == READ_WRITE }
 }
