+++ test/DictionarySpec.scala
@@ -14,37 +14,31 @@ import controllers.DictionaryApp
 
 class DictionarySpec extends PlaySpec with Results with OneAppPerSuite {
 
-  class TestController() extends Controller with DictionaryApp
-
-  val controller = new TestController()
-  import controller.{ add, search, Dictionary, Users }
-
   "add service" should {
 
     "allow adding new words if the user is empowered to do so" in {
-      val word = "new"
       val request = FakeRequest(
-	POST, 
-	"/", 
-	FakeHeaders(Seq(("user", Seq("mr_proper")))),
-	(word, "a new definition"))
-      val result = add apply request
+                    	POST, 
+                    	"/", 
+                    	FakeHeaders(Seq(("user", Seq("mr_proper")))),
+    	                ("new", "a new definition"))
+      val result = DictionaryApp.add(request) // así se ve mejor que el controlador es una función
       status(result) mustEqual OK
     }
 
     "fail if the user is not empowered to do so" in {
-      val word = "new"
+      // val word = "new" // si la definition no se incluye en una variable, no sé por qué sí la palabra
       val request = FakeRequest(
-	POST, 
-	"/", 
-        FakeHeaders(Seq(("user", Seq("don_limpio")))),
-	(word, "a brand new definition"))
-      val result: Future[Result] = add apply request
+                    	POST, 
+                    	"/", 
+                      FakeHeaders(Seq(("user", Seq("don_limpio")))),
+	                    ("new", "a brand new definition"))
+      val result: Future[Result] = DictionaryApp.add(request)
       status(result) mustEqual FORBIDDEN
       contentAsString(result) mustEqual "You are not allowed to write"
 
       val request2 = request.withHeaders(("user", "wipp_express"))
-      val result2 = add apply request2
+      val result2 = DictionaryApp.add(request2)
       status(result2) mustEqual FORBIDDEN
       contentAsString(result2) mustEqual "You are not allowed to write"
     }
@@ -56,8 +50,8 @@ class DictionarySpec extends PlaySpec with Results with OneAppPerSuite {
       val word = "known"
       Dictionary.set(word -> "a well known word")
       val request = FakeRequest(GET, s"/$word")
-	.withHeaders(("user" -> "don_limpio"))
-      val result: Future[Result] = search(word) apply request
+	                   .withHeaders(("user" -> "don_limpio"))
+      val result: Future[Result] = DictionaryApp.search(word)(request)
       status(result) mustEqual OK
       contentAsString(result) mustEqual "a well known word"
     }
@@ -65,8 +59,8 @@ class DictionarySpec extends PlaySpec with Results with OneAppPerSuite {
     "not find a non-existing word" in {
       val word = "unknown"
       val request = FakeRequest(GET, s"/$word")
-	.withHeaders(("user" -> "don_limpio"))
-      val result: Future[Result] = search(word) apply request
+                      .withHeaders(("user" -> "don_limpio"))
+      val result: Future[Result] = DictionaryApp.search(word)(request)
       status(result) mustEqual NOT_FOUND
       contentAsString(result) mustEqual s"The word '$word' does not exist"
     }
